+++ packages/plugin-workflow-approval/src/client/usage/h5/component/ApprovalReachDataType.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { observer } from '@tachybase/schema';
 import { dayjs } from '@tachybase/utils/client';
