+++ packages/plugin-workflow-approval/src/client/usage/h5/context/SchemaComponent.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { SchemaComponentContext, useSchemaComponentContext } from '@tachybase/client';
 
 export function SchemaComponentContextProvider({ designable, children }) {
