+++ packages/plugin-workflow-approval/src/client/usage/h5/context/WithActionEnabled.tsx
@@ -1,3 +1,5 @@
 import { createContext } from 'react';
 
-export const ContextWithActionEnabled = createContext({ actionEnabled: false });
+export const ContextWithActionEnabled = createContext({
+  actionEnabled: false,
+});
