+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/component/ApprovalItem.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { useAPIClient, useCollectionManager, useCompile, useCurrentUserContext } from '@tachybase/client';
 import { observer } from '@tachybase/schema';
 import { dayjs } from '@tachybase/utils/client';
@@ -8,8 +8,8 @@ import { Empty, List, Space, Tag } from 'antd-mobile';
 import { useNavigate } from 'react-router-dom';
 
 import { approvalStatusEnums } from '../../../../common/constants/approval-initiation-status-options';
+import { useTranslation } from '../../../../locale';
 import { ApprovalPriorityType } from '../../constants';
-import { useTranslation } from '../../locale';
 import { InitiationsBlockContext } from '../InitiationsBlock';
 
 export const ApprovalItem = observer((props) => {
