+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/component/ApprovalStatus.tsx
@@ -1,10 +1,10 @@
-import React, { useState } from 'react';
-import { observer, useFieldSchema } from '@tachybase/schema';
+import { useState } from 'react';
+import { observer } from '@tachybase/schema';
 
 import { Picker, Space } from 'antd-mobile';
 import { DownOutline } from 'antd-mobile-icons';
 
-import { useTranslation } from '../../locale';
+import { useTranslation } from '../../../../locale';
 
 export const ApprovalStatus = observer((props) => {
   const { changeFilter, filter } = props as any;
