+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/component/ViewActionInitiationsContent.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useEffect, useState } from 'react';
+import { createContext, useContext, useEffect, useState } from 'react';
 import {
   CollectionProvider_deprecated,
   MobileProvider,
@@ -12,15 +12,14 @@ import {
 import { NavBar } from 'antd-mobile';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { useTranslation } from '../../locale';
-
-import '../../style/style.css';
-
 import { APPROVAL_INITIATION_STATUS } from '../../../../common/constants/approval-initiation-status';
+import { useTranslation } from '../../../../locale';
 import { useActionReminder } from '../hook/useActionReminder';
 import { useActionResubmit } from '../hook/useActionResubmit';
 import { useCreateSubmit } from '../hook/useCreateSubmit';
 import { ProviderActionReminder } from '../provider/ActionReminder.provider';
+
+import '../../style/style.css';
 
 export const ViewActionInitiationsContent = () => {
   const params = useParams();
