+++ packages/plugin-workflow-approval/src/common/constants.ts
@@ -3,7 +3,7 @@ export const NOTICE_INSTRUCTION_NAMESPACE = 'notice';
 export const COLLECTION_NOTICE_NAME = 'workflowNotice';
 export const COLLECTION_NAME_APPROVAL_CARBON_COPY = 'approvalCarbonCopy';
 export const COLLECTION_WORKFLOWS_NAME = 'workflows';
-export const NAMESPACE = '@tachybase/plugin-workflow-approval';
+export const NAMESPACE = 'workflow-approval';
 
 export const PLUGIN_NAME_APPROVAL = 'approval';
 export const INSTRUCTION_TYPE_NAME_APPROVAL = 'approval';
